refactor: 이미지 관련 컴포넌트 분리
