+++ frontend/src/components/PostForm/components/ImageUpload/index.styles.ts
@@ -1,0 +1,16 @@
+import Camera from '@/assets/images/camera.svg';
+
+import styled from '@emotion/styled';
+
+export const ImageInput = styled.input`
+  display: none;
+`;
+
+export const ImageUploadButton = styled.label`
+  cursor: pointer;
+  padding-left: 10px;
+  padding-right: 10px;
+  margin-right: -10px;
+`;
+
+export const CameraIcon = styled(Camera)``;
