+++ frontend/src/components/PostForm/index.styles.ts
@@ -1,9 +1,6 @@
 import CheckBoxComponent from '@/components/CheckBox';
 
-import Camera from '@/assets/images/camera.svg';
-
 import { invalidInputAnimation } from '@/style/GlobalStyle';
-import { keyframes } from '@emotion/react';
 import styled from '@emotion/styled';
 
 export const Container = styled.form`
@@ -77,67 +74,9 @@ export const SubmitButton = styled.button`
 
 export const CheckBox = styled(CheckBoxComponent)``;
 
-export const ImagePreview = styled.div`
-  width: 100%;
-  height: 100px;
-  display: flex;
-`;
-
-export const Image = styled.img`
-  width: 100px;
-  height: 100px;
-  border-radius: 5px;
-`;
-
-export const ImageName = styled.p`
-  font-size: 14px;
-  color: ${props => props.theme.colors.gray_200};
-  padding-left: 20px;
-  line-height: 20px;
-  user-select: none;
-`;
-
-export const ImageUploadLoading = styled.div`
-  width: 100%;
-  height: 100px;
-  color: ${props => props.theme.colors.gray_200};
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  font-family: 'BMHANNAAir', 'Noto Sans KR';
-  font-size: 18px;
-`;
-
-export const rotate = keyframes`
-  0% {
-    transform: rotate(0deg);
-  } 
-  100% {
-    transform: rotate(360deg);
-  }
-`;
-
-export const LoadingIcon = styled.span`
-  animation: ${rotate} 1.5s linear 0s infinite;
-  margin-left: 10px;
-`;
-
 export const PostController = styled.div`
   width: 100%;
   display: flex;
   align-items: center;
   justify-content: space-between;
-`;
-
-export const CameraIcon = styled(Camera)``;
-
-export const ImageInput = styled.input`
-  display: none;
-`;
-
-export const ImageUploadButton = styled.label`
-  cursor: pointer;
-  padding-left: 10px;
-  padding-right: 10px;
-  margin-right: -10px;
 `;
