+++ frontend/src/types/post.d.ts
@@ -20,3 +20,9 @@ interface Post {
   blocked: boolean;
   imageName: string;
 }
+
+interface Image {
+  file?: File;
+  src?: string;
+  path: string;
+}
